Comment out backend test that uses requests
We can't actually make calls to the frontend (see errors in build 26) without doing a whole bunch of other work (this is what Professor da Silva described to me anyway and I think I forwarded all of the info over slack). Anyway, just commenting out for now until we decide what else to do with it.